+++ lib/event_store/config.ex
@@ -84,7 +84,8 @@ defmodule EventStore.Config do
     :pool_size,
     :queue_target,
     :queue_interval,
-    :socket_options
+    :socket_options,
+    :parameters
   ]
 
   def default_postgrex_opts(config) do